pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "binghuis", email = "soongbinghui@gmail.com" }]
dependencies = [
    "fastapi>=0.110.2",
    "uvicorn[standard]>=0.29.0",
    "langchain-community>=0.0.34",
    "langchain-openai>=0.1.4",
    "pydantic-settings>=2.2.1",
    "langchain-text-splitters>=0.0.1",
    "langchain>=0.1.19",
    "bs4>=0.0.2",
    "langchain-chroma>=0.1.0",
    "types-beautifulsoup4>=4.12.0.20240504",
    "langchainhub>=0.1.15",
    "lxml>=5.2.1",
    "html5lib>=1.1",
    "azure-ai-documentintelligence>=1.0.0b3",
]
requires-python = "<3.13,>=3.12"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "uvicorn src.server.main:app --reload"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
